First commit, refactoring of code 	modified:   .gitignore 	renamed:    elo_calc.py -> src/elo_calc.py 	renamed:    process.py -> src/process.py 	renamed:    season_creator.py -> src/season_creator.py 	new file:   test/Gen9OU-2023-04-26-wmossdeepwalrein-chocoq_1_1.html
