+++ .gitignore
@@ -1,3 +1,4 @@
 env/
 __pycache__/
 records/
+.history/
